requirements.txt
kafka-python>=2.0.2
pydantic>=1.10.15

oracledb>=2.2.0

fastapi>=0.112.0
uvicorn>=0.30.0
openai>=1.40.0
